refactoring, bot votes now count
