+++ .prettierrc.json
@@ -1,0 +1,10 @@
+{
+  "printWidth": 110,
+  "useTabs": false,
+  "tabWidth": 2,
+  "singleQuote": false,
+  "semi": true,
+  "trailingComma": "es5",
+  "arrowParens": "always",
+  "endOfLine": "lf"
+}
