+++ src/bot/commands/theme.ts
@@ -1,11 +1,28 @@
-import { ChatInputCommandInteraction, MessageFlags, EmbedBuilder, ButtonBuilder, ButtonStyle, ActionRowBuilder, ButtonInteraction, ModalBuilder, TextInputBuilder, TextInputStyle, ModalSubmitInteraction, LabelBuilder } from "discord.js";
+import {
+  ChatInputCommandInteraction,
+  MessageFlags,
+  EmbedBuilder,
+  ButtonBuilder,
+  ButtonStyle,
+  ActionRowBuilder,
+  ButtonInteraction,
+  ModalBuilder,
+  TextInputBuilder,
+  TextInputStyle,
+  ModalSubmitInteraction,
+  LabelBuilder,
+} from "discord.js";
 import { getOrCreateUser, updateUser, getUser } from "../../database";
 
 // Show modal for /daily-theme command
 async function handleDailyThemeCommand(interaction: ChatInputCommandInteraction) {
   try {
     const guild = interaction.guild;
-    if (!guild) return interaction.reply({ content: "Guild not found.", flags: MessageFlags.Ephemeral });
+    if (!guild)
+      return interaction.reply({
+        content: "Guild not found.",
+        flags: MessageFlags.Ephemeral,
+      });
 
     const userId = interaction.user.id;
     const guildId = guild.id;
@@ -16,10 +33,19 @@ async function handleDailyThemeCommand(interaction: ChatInputCommandInteraction)
     const desc = user?.themeDescription || "";
 
     const { embed, row } = buildSavedThemeEmbed(title, desc);
-    await interaction.reply({ embeds: [embed], components: [row], flags: MessageFlags.Ephemeral });
+    await interaction.reply({
+      embeds: [embed],
+      components: [row],
+      flags: MessageFlags.Ephemeral,
+    });
   } catch (e) {
     console.error("Error showing daily theme status:", e);
-    try { await interaction.reply({ content: "Failed to show theme status.", flags: MessageFlags.Ephemeral }); } catch { }
+    try {
+      await interaction.reply({
+        content: "Failed to show theme status.",
+        flags: MessageFlags.Ephemeral,
+      });
+    } catch {}
   }
 }
 
@@ -35,9 +61,7 @@ async function handleLaunchDailyThemeModal(interaction: ButtonInteraction) {
       .setRequired(true)
       .setMaxLength(100);
 
-    const titleLabel = new LabelBuilder()
-      .setLabel("Daily theme title")
-      .setTextInputComponent(titleInput);
+    const titleLabel = new LabelBuilder().setLabel("Daily theme title").setTextInputComponent(titleInput);
 
     // Description
     const descInput = new TextInputBuilder()
@@ -46,23 +70,30 @@ async function handleLaunchDailyThemeModal(interaction: ButtonInteraction) {
       .setRequired(false)
       .setMaxLength(400);
 
-    const descLabel = new LabelBuilder()
-      .setLabel("More details (optional)")
-      .setTextInputComponent(descInput);
+    const descLabel = new LabelBuilder().setLabel("More details (optional)").setTextInputComponent(descInput);
 
     modal.addLabelComponents(titleLabel, descLabel);
 
     await interaction.showModal(modal);
   } catch (e) {
     console.error("Error launching daily theme modal:", e);
-    try { await interaction.reply({ content: "Failed to open theme form.", flags: MessageFlags.Ephemeral }); } catch { }
+    try {
+      await interaction.reply({
+        content: "Failed to open theme form.",
+        flags: MessageFlags.Ephemeral,
+      });
+    } catch {}
   }
 }
 
 async function handleClearDailyThemeButton(interaction: ButtonInteraction) {
   try {
     const guild = interaction.guild;
-    if (!guild) return interaction.reply({ content: "Guild not found.", flags: MessageFlags.Ephemeral });
+    if (!guild)
+      return interaction.reply({
+        content: "Guild not found.",
+        flags: MessageFlags.Ephemeral,
+      });
 
     const userId = interaction.user.id;
     const guildId = guild.id;
@@ -70,7 +101,11 @@ async function handleClearDailyThemeButton(interaction: ButtonInteraction) {
 
     // Ensure user exists
     getOrCreateUser(userId, guildId, username);
-    updateUser(userId, guildId, { themeTitle: null, themeDescription: null, themeTimestampUTC: null });
+    updateUser(userId, guildId, {
+      themeTitle: null,
+      themeDescription: null,
+      themeTimestampUTC: null,
+    });
 
     const user = getUser(userId, guildId);
     const title = user?.themeTitle || "(no saved theme)";
@@ -80,7 +115,13 @@ async function handleClearDailyThemeButton(interaction: ButtonInteraction) {
     await interaction.update({ embeds: [embed], components: [row] });
   } catch (e) {
     console.error("Error clearing daily theme:", e);
-    try { if (!interaction.replied && !interaction.deferred) await interaction.reply({ content: "Failed to clear theme.", flags: MessageFlags.Ephemeral }); } catch { }
+    try {
+      if (!interaction.replied && !interaction.deferred)
+        await interaction.reply({
+          content: "Failed to clear theme.",
+          flags: MessageFlags.Ephemeral,
+        });
+    } catch {}
   }
 }
 
@@ -88,7 +129,11 @@ async function handleClearDailyThemeButton(interaction: ButtonInteraction) {
 async function handleDailyThemeModalSubmit(interaction: ModalSubmitInteraction) {
   try {
     const guild = interaction.guild;
-    if (!guild) return interaction.reply({ content: "Guild not found.", flags: MessageFlags.Ephemeral });
+    if (!guild)
+      return interaction.reply({
+        content: "Guild not found.",
+        flags: MessageFlags.Ephemeral,
+      });
 
     const userId = interaction.user.id;
     const guildId = guild.id;
@@ -116,11 +161,20 @@ async function handleDailyThemeModalSubmit(interaction: ModalSubmitInteraction) 
       await interaction.editReply({ embeds: [embed], components: [row] });
     } catch (e) {
       console.log("Failed to edit reply, sending new reply instead", e);
-      await interaction.reply({ embeds: [embed], components: [row], flags: MessageFlags.Ephemeral });
+      await interaction.reply({
+        embeds: [embed],
+        components: [row],
+        flags: MessageFlags.Ephemeral,
+      });
     }
   } catch (e) {
     console.error("Error handling daily theme submit:", e);
-    try { await interaction.reply({ content: "Failed to save theme.", flags: MessageFlags.Ephemeral }); } catch { }
+    try {
+      await interaction.reply({
+        content: "Failed to save theme.",
+        flags: MessageFlags.Ephemeral,
+      });
+    } catch {}
   }
 }
 
@@ -130,12 +184,15 @@ function buildSavedThemeEmbed(title: string, desc: string) {
     .setTitle("Daily Drawing Theme")
     .setDescription("Save a daily drawing theme that will automatically be posted if you win")
     .setColor(0x0099ff)
-    .addFields(
-      { name: "Title", value: title },
-      { name: "Description", value: desc || "(none)" }
-    );
-  const updateButton = new ButtonBuilder().setCustomId("daily-theme-update").setLabel("Update Theme").setStyle(ButtonStyle.Primary);
-  const clearButton = new ButtonBuilder().setCustomId("daily-theme-clear").setLabel("Clear Theme").setStyle(ButtonStyle.Danger);
+    .addFields({ name: "Title", value: title }, { name: "Description", value: desc || "(none)" });
+  const updateButton = new ButtonBuilder()
+    .setCustomId("daily-theme-update")
+    .setLabel("Update Theme")
+    .setStyle(ButtonStyle.Primary);
+  const clearButton = new ButtonBuilder()
+    .setCustomId("daily-theme-clear")
+    .setLabel("Clear Theme")
+    .setStyle(ButtonStyle.Danger);
   const row = new ActionRowBuilder<ButtonBuilder>().addComponents(updateButton, clearButton);
   return { embed, row };
 }
