+++ src/reactionCheck.ts
@@ -1,6 +1,5 @@
-import { Message, GuildMember, PermissionsBitField } from "discord.js";
-
-const modRoles: string[] = process.env.MOD_ROLES ? process.env.MOD_ROLES.split(",").map((r) => r.trim()) : [];
+import { Message } from "discord.js";
+import { userHasModPermission } from "./util";
 
 /**
  * Check if a message is marked as overtime (timer emoji reacted by the author).
@@ -43,7 +42,7 @@ async function isMarkedOvertime(reply: Message<true>): Promise<boolean> {
 }
 
 /**
- * Count unique users (excluding bots and the author) who reacted with fire emoji.
+ * Count unique users (excluding the author) who reacted with fire emoji.
  * Returns the count of unique fire reactors.
  */
 async function countFireReactors(reply: Message<true>): Promise<number> {
@@ -55,7 +54,7 @@ async function countFireReactors(reply: Message<true>): Promise<number> {
         if (emojiName !== "🔥" && emojiName?.toLowerCase() !== "fire") continue;
         const users = await reaction.users.fetch();
         users.forEach((user) => {
-          if (!user.bot && user.id !== (reply.author?.id ?? "")) uniqueUserIds.add(user.id);
+          if (user.id !== (reply.author?.id ?? "")) uniqueUserIds.add(user.id);
         });
       } catch (e) {
         // Ignore fetch errors for individual reactions
@@ -68,30 +67,6 @@ async function countFireReactors(reply: Message<true>): Promise<number> {
   }
 }
 
-/**
- * Check if a user has mod permissions (either specified mod role or kick permission).
- */
-function userHasModPermission(member: GuildMember | null): boolean {
-  if (!member) return false;
 
-  // Check for kick permission
-  const canKick = member.permissions?.has?.(PermissionsBitField.Flags?.KickMembers ?? 0);
-  if (canKick) return true;
 
-  // Check for admin
-  const isAdmin = member.permissions?.has?.(PermissionsBitField.Flags?.Administrator ?? 0);
-  if (isAdmin) return true;
-
-  // Check for mod roles
-  if (modRoles.length > 0) {
-    for (const role of member.roles.cache.values()) {
-      if (modRoles.includes(role.id) || modRoles.includes(role.name)) {
-        return true;
-      }
-    }
-  }
-
-  return false;
-}
-
-export { isMarkedOvertime, countFireReactors, userHasModPermission };
+export { isMarkedOvertime, countFireReactors };
